+++ app/Module.php
@@ -4,7 +4,7 @@ namespace GeeksAreForLife\TrelloMate;
 
 abstract class Module
 {
-    abstract public function __construct(&$trello, &$config);
+    abstract public function __construct(&$trello, &$output, &$config);
 
     abstract public function getCommands();
