+++ trellomate
@@ -5,5 +5,4 @@ require __DIR__ . '/vendor/autoload.php';
 use GeeksAreForLife\TrelloMate\TrelloMate;
 
 // hand off to the trellomate object
-$debug = true;
-new TrelloMate($debug);
+new TrelloMate();
